+++ packages/next-docs-ui/src/nav.tsx
@@ -1,5 +1,2 @@
 'use client'
 export * from './components/nav'
-export { SidebarContext } from 'next-docs-zeta/sidebar'
-export { SearchContext } from './contexts/search'
-export { SidebarContext as DesktopSidebarContext } from './contexts/sidebar'